Several ops/txn in insert and update
The insertion and update tests were modified
in order to allow several operations per transaction.

A new flag, -a, was added to specify the number of
operations to perform per transaction.

The flag -n still means the number of transactions performed,
so: "insert -n 100 -a 100 -w 1" will insert 10000 rows. This
behavior was choosen in order to reduce the number of
modifications to Benchmark, and don't affect that much previous
tests measures.

Change-Id: I1dc0e108b55b18b2dbfbf49784789ca2972f8e63
Signed-off-by: Esteban Rodriguez Betancourt <estebarb@hpe.com>